# server/openapi.yaml
openapi: 3.0.3
info:
  title: Campus Study Buddy API
  description: API for Campus Study Buddy - A platform for students to find study partners and form study groups
  version: 1.0.0
  contact:
    name: Campus Study Buddy Team
    email: support@campusstudybuddy.com
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.campusstudybuddy.com
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token
  schemas:
...
    StudyGroup:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        module:
          type: string
        maxMembers:
          type: integer
          minimum: 1
        currentMembers:
          type: integer
          minimum: 0
        members:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              role:
                type: string
                enum: [admin, member]
        schedule:
          type: array
          items:
            type: string
        location:
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
    Message:
      type: object
      properties:
        id:
          type: string
        groupId:
          type: string
        userId:
          type: string
        userName:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...
paths:
...
  /api/groups:
    get:
      summary: Get study groups
      description: Retrieve a list of study groups, optionally filtered by module
      tags:
        - Groups
      security:
        - BearerAuth: []
      parameters:
        - name: module
          in: query
          description: Filter groups by module code
          required: false
          schema:
            type: string
            example: COMS3011
      responses:
        '200':
          description: List of study groups
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StudyGroup'
                  count:
                    type: integer
        '401':
          description: Unauthorized - Invalid or missing token
        '500':
          description: Internal server error

    post:
      summary: Create a study group
      description: Create a new study group
      tags:
        - Groups
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - module
              properties:
                name:
                  type: string
                  description: Name of the study group
                  minLength: 1
                  maxLength: 100
                description:
                  type: string
                  description: Description of the study group
                  maxLength: 500
                module:
                  type: string
                  description: Module code
                  minLength: 1
                  maxLength: 20
                maxMembers:
                  type: integer
                  minimum: 2
                  maximum: 50
                  default: 8
                  description: Maximum number of members
                schedule:
                  type: array
                  items:
                    type: string
                  description: Study schedule
                location:
                  type: string
                  description: Study location
                  maxLength: 200
      responses:
        '201':
          description: Study group created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/StudyGroup'
                  message:
                    type: string
        '400':
          description: Bad request - Validation error
        '401':
          description: Unauthorized - Invalid or missing token
        '500':
          description: Internal server error
  /api/groups/{id}:
    get:
      summary: Get a specific study group
      description: Retrieve details of a specific study group by ID
      tags:
        - Groups
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: string
      responses:
        '200':
          description: Group details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/StudyGroup'
        '400':
          description: Bad request - Validation error
        '401':
          description: Unauthorized - Invalid or missing token
        '404':
          description: Group not found
        '500':
          description: Internal server error
  /api/groups/{id}/join:
    post:
      summary: Join a study group
      description: Join an existing study group
      tags:
        - Groups
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: string
      responses:
        '200':
          description: Successfully joined the group
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/StudyGroup'
                  message:
                    type: string
        '400':
          description: Bad request - Already a member or group is full
        '401':
          description: Unauthorized - Invalid or missing token
        '404':
          description: Group not found
        '500':
          description: Internal server error
  /api/groups/{id}/message:
    post:
      summary: Send a message to a study group
      description: Send a message to a study group (must be a member)
      tags:
        - Groups
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  description: Message content
                  minLength: 1
                  maxLength: 1000
      responses:
        '201':
          description: Message sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Message'
                  message:
                    type: string
        '400':
          description: Bad request - Validation error
        '401':
          description: Unauthorized - Invalid or missing token
        '403':
          description: Forbidden - Not a member of the group
        '404':
          description: Group not found
        '500':
          description: Internal server error
  /api/groups/{id}/messages:
    get:
      summary: Get messages from a study group
      description: Retrieve messages from a study group (must be a member)
      tags:
        - Groups
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of messages to return
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
                  count:
                    type: integer
                  total:
                    type: integer
        '400':
          description: Bad request - Validation error
        '401':
          description: Unauthorized - Invalid or missing token
        '403':
          description: Forbidden - Not a member of the group
        '404':
          description: Group not found
        '500':
          description: Internal server error
...